requirements.txt
langchain
langchain_community
langchain-groq
chromadb
streamlit
time
zipfile
pandas
python-dotenv
os